% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompSiteCDF.R
\name{CompSiteCDF}
\alias{CompSiteCDF}
\title{CompSiteCDF, compare CDFs of sites}
\usage{
CompSiteCDF(df.input, dir.input, dir.output, ParamName.xlab)
}
\arguments{
\item{df.input}{Input data as CSV.  Needs 6 columns (SampleID and up to 5 SiteIDs).}

\item{dir.input}{Directory where data file is located.}

\item{dir.output}{Directory where PDF file is to be saved.}

\item{ParamName.xlab}{Parameter name for x-axis on plots}
}
\value{
Returns a PDF of CDFs.
}
\description{
Takes as an input a data frame with date and up to 5 columns of parameter data.
Column names are SiteIDs and values are daily means for some measurement.
}
\details{
CDFs are generate for year, season, and year/season and saved to a PDF
}
\examples{
# load bio data
df.data <- data_CompSiteCDF
dim(df.data)
View(df.data)

# subsample
mySize <- 200
Seed.MS <- 18171210
bugs.mysize <- rarify(inbug=DF.biodata, sample.ID="SampRep",abund="Count",subsiz=mySize, mySeed=Seed.MS)
dim(bugs.mysize)
View(bugs.mysize)
# save the data
write.table(bugs.mysize,paste("bugs",mySize,"txt",sep="."),sep="\\t")
}
\keyword{CDF,}
\keyword{comparison}
\keyword{continuous}
\keyword{data,}
\keyword{site}
