% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompSiteCDF.updated.R
\name{CompSiteCDF.updated}
\alias{CompSiteCDF.updated}
\title{CompSiteCDF.updated, compare CDFs of sites
this R function is adopted from "CompSiteCDF"
from https://github.com/leppott/ContDataQC/blob/main/R/CompSiteCDF.R}
\usage{
CompSiteCDF.updated(
  file.input = NULL,
  dir.input = getwd(),
  dir.output = getwd(),
  ParamName.xlab = " ",
  Plot.title = " ",
  Plot.type = "Density",
  Plot.season = "Summer",
  hist.columnName = NULL,
  df.input = NULL
)
}
\description{
CompSiteCDF.updated, compare CDFs of sites
this R function is adopted from "CompSiteCDF"
from https://github.com/leppott/ContDataQC/blob/main/R/CompSiteCDF.R
}
\examples{
CompSiteCDF.updated(ParamName.xlab = myXlab, df.input=myDF)

}
\keyword{internal}
