% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_bio2rarify}
\alias{data_bio2rarify}
\title{data_bio2rarify}
\format{a data frame with 74 observations and 3 variables:}
\usage{
data_bio2rarify
}
\description{
Test data for rarify().  Includes biological data for 2 samples that are both over the 240 total count limit.
}
\keyword{datasets}
