% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_CompSiteCDF}
\alias{data_CompSiteCDF}
\title{data_CompSiteCDF}
\format{a data frame with 349 observations and 6 variables:}
\usage{
data_CompSiteCDF
}
\description{
Test data for CompSiteCDF().  Includes daily mean water temp (C) data for 5.
}
\keyword{datasets}
