% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{DATA_period_test2_Aw_20130101_20141231}
\alias{DATA_period_test2_Aw_20130101_20141231}
\title{DATA_period_test2_Aw_20130101_20141231}
\format{a data frame with 9699 observations and 72 variables:}
\usage{
DATA_period_test2_Aw_20130101_20141231
}
\description{
Test data for PeriodStats().  Includes air and water data at different start times.
}
\keyword{datasets}
