% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_raw_ECO66G12_AW_20160128_20160418}
\alias{data_raw_ECO66G12_AW_20160128_20160418}
\title{data_raw_ECO66G12_AW_20160128_20160418}
\format{
a data frame with 771 observations and 1 variables:
}
\usage{
data_raw_ECO66G12_AW_20160128_20160418
}
\description{
Test data for formatHobo().  Example output from HOBOware.
}
\keyword{datasets}
