% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_raw_test2_AW_20130426_20130725}
\alias{data_raw_test2_AW_20130426_20130725}
\title{data_raw_test2_AW_20130426_20130725}
\format{
a data frame with 4320 observations and 11 variables:
}
\usage{
data_raw_test2_AW_20130426_20130725
}
\description{
Test data
}
\keyword{datasets}
