% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_raw_test2_AW_20140901_20140930}
\alias{data_raw_test2_AW_20140901_20140930}
\title{data_raw_test2_AW_20140901_20140930}
\format{a data frame with 1440 observations and 18 variables:}
\usage{
data_raw_test2_AW_20140901_20140930
}
\description{
Test data.
}
\keyword{datasets}
