% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_raw_test4_AW_20160418_20160726}
\alias{data_raw_test4_AW_20160418_20160726}
\title{data_raw_test4_AW_20160418_20160726}
\format{
a data frame with 9508 observations and 8 variables:
}
\usage{
data_raw_test4_AW_20160418_20160726
}
\description{
Test data.  Includes air and water data at different start times.
}
\keyword{datasets}
