% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.DateFormat.R
\name{fun.dt.Type2}
\alias{fun.dt.Type2}
\title{Date Format (function)}
\usage{
fun.dt.Type2(fun.dt, fun2.dt.Type)
}
\arguments{
\item{fun.dt}{Date, Time, or DateTime data}

\item{fun.dt2.Type}{type of input; date, time, or date}
}
\value{
Returns a text string representing the date/time format of the input fun.dt.  Wrapped in function fun.DateTimeFormat().
}
\description{
Date Format (function)
}
\examples{
#Not intended to be accessed indepedant of function "fun.DateTimeFormat()".
fun.dt.Type2(Sys.Date(),"date")
}
\keyword{date,}
\keyword{datetime,}
\keyword{format}
\keyword{time,}

