% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.GageData.R
\name{fun.GageData}
\alias{fun.GageData}
\title{Download USGS Gage Data}
\usage{
fun.GageData(fun.myData.SiteID, fun.myData.Type = "Gage",
  fun.myData.DateRange.Start, fun.myData.DateRange.End,
  fun.myDir.export = getwd(), myTZ = ContData.env$myTZ)
}
\arguments{
\item{fun.myData.SiteID}{Station/SiteID (can be a vector).}

\item{fun.myData.Type}{data type is "Gage".}

\item{fun.myData.DateRange.Start}{Start date for requested data. Format = YYYY-MM-DD.}

\item{fun.myData.DateRange.End}{End date for requested data. Format = YYYY-MM-DD.}

\item{fun.myDir.export}{SDirectory for export data.  Default is current working directory.}

\item{myTZ}{Timezone for requested gage.  Default is in env.UserDefinedValues.R.  Can also be set with Sys.timezone().}
}
\value{
Returns a csv file to specified directory with the requested daily mean data.  During the data retrieval a summary is output to the console.
}
\description{
This a wrapper function for the package dataRetrieval to get USGS data.  Daily means are the default data type.
}
\examples{
#Not intended to be accessed indepedant of function ContDataQC().
fun.GageData("01187300","Gage","2013-01-01","2013-12-31",getwd(),"","")
# with optional variable left blank
fun.GageData("01187300","Gage","2013-01-01","2013-12-31")
}
\keyword{USGS,}
\keyword{continuous}
\keyword{data,}
\keyword{dataRetrieval}
\keyword{gage,}

