% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.QC.R
\name{fun.QC}
\alias{fun.QC}
\title{Quality Control}
\usage{
fun.QC(fun.myData.SiteID, fun.myData.Type = "QC", fun.myData.DateRange.Start,
  fun.myData.DateRange.End, fun.myDir.BASE = getwd(),
  fun.myDir.SUB.import = "", fun.myDir.SUB.export = "")
}
\arguments{
\item{fun.myData.SiteID}{Station/SiteID.}

\item{fun.myData.Type}{data type is "QC".}

\item{fun.myData.DateRange.Start}{Start date for requested data. Format = YYYY-MM-DD.}

\item{fun.myData.DateRange.End}{End date for requested data. Format = YYYY-MM-DD.}

\item{fun.myDir.BASE}{Root directory for data.  If blank will use current working directory.}

\item{fun.myDir.SUB.import}{Subdirectory for import data.  If blank will use root directory.}

\item{fun.myDir.SUB.export}{Subdirectory for export data.  If blank will use root directory.}
}
\value{
Returns a csv file to specified directory with QC flags.
}
\description{
Subfunction for performing QC.  Needs to be called from ContDataQC().
Requires zoo().
}
\examples{
#Not intended to be accessed indepedant of function ContDataQC().
}
\keyword{continuous}
\keyword{control}
\keyword{data,}
\keyword{qc,}
\keyword{quality}

