% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.Report.File.R
\name{fun.Report.File}
\alias{fun.Report.File}
\title{Generate Reports, File}
\usage{
fun.Report.File(fun.myFile, fun.myDir.import = getwd(),
  fun.myDir.export = getwd(), fun.myFile.Prefix, fun.myReport.format)
}
\arguments{
\item{fun.myFile}{Single file name.}

\item{fun.myDir.import}{Directory for import data.  Default is current working directory.}

\item{fun.myDir.export}{Directory for export data.  Default is current working directory.}

\item{fun.myFile.Prefix}{Valid prefixes are "QC", "DATA", or "STATS".  This determines the RMD to use for the outpu.}

\item{fun.myReport.format}{Report format (docx or html).  Default is specified in config.R (docx).}
}
\value{
Creates a Word file in the specified directory.
}
\description{
Subfunction for generating Word reports.  Needs to be called from ContDataQC().
Requires knitr() and pandoc.
Different reports (in Word) are created using the RMD files in this package depending on the input fun.myFile.Prefix.
Values are "QC", "DATA", or "STATS".
The 'file' version works on a single (or vector) of files.
}
\examples{
#Not intended to be accessed indepedant of function ContDataQC().
}
\keyword{continuous}
\keyword{data,}
\keyword{report}
