% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.Report.R
\name{fun.Report}
\alias{fun.Report}
\title{Generate Reports}
\usage{
fun.Report(fun.myData.SiteID, fun.myData.Type, fun.myData.DateRange.Start,
  fun.myData.DateRange.End, fun.myDir.BASE = getwd(),
  fun.myDir.SUB.import = "", fun.myDir.SUB.export = "", fun.myFile.Prefix)
}
\arguments{
\item{fun.myData.SiteID}{Station/SiteID.}

\item{fun.myData.Type}{data type; c("Air","Water","AW","Gage","AWG","AG","WG")}

\item{fun.myData.DateRange.Start}{Start date for requested data. Format = YYYY-MM-DD.}

\item{fun.myData.DateRange.End}{End date for requested data. Format = YYYY-MM-DD.}

\item{fun.myDir.BASE}{Root directory for data.  If blank will use current working directory.}

\item{fun.myDir.SUB.import}{Subdirectory for import data.  If blank will use root directory.}

\item{fun.myDir.SUB.export}{Subdirectory for export data.  If blank will use root directory.}

\item{fun.myFile.Prefix}{Valid prefixes are "QC", "DATA", or "STATS".  This determines the RMD to use for the outpu.}
}
\value{
Creates a Word file in the specified directory.
}
\description{
Subfunction for generating Word reports.  Needs to be called from ContDataQC().
Requires knitr() and pandoc.
Different reports (in Word) are created using the RMD files in this package depending on the input fun.myFile.Prefix.
Values are "QC", "DATA", or "STATS".
}
\examples{
#Not intended to be accessed indepedant of function ContDataQC().
}
\keyword{continuous}
\keyword{data,}
\keyword{report}

