% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.Stats.File.R
\name{fun.Stats.File}
\alias{fun.Stats.File}
\title{Statistical Summary, File}
\usage{
fun.Stats.File(fun.myFile, fun.myDir.import = getwd(),
  fun.myDir.export = getwd(), fun.myReport.format)
}
\arguments{
\item{fun.myFile}{Single (or vector of) file name(s).}

\item{fun.myDir.import}{Directory for import data.  Default is current working directory.}

\item{fun.myDir.export}{Directory for export data.  Default is current working directory.}

\item{fun.myReport.format}{Report format (docx or html).  Default is specified in config.R (docx).}
}
\value{
Returns a csv into the specified export directory with additional columns for calculated statistics.  And a PDF of summary plots by different time periods.
}
\description{
Subfunction for generating statistical summaries.  Needs to be called from ContDataQC().
Requires doBy() and survival() [required by doBy]
Calculates statistics on input data and saves to a new csv.  Plots are saved as  PDF.
The 'file' version works on a single (or vector) of files.
}
\examples{
myFile <- "QC_test2_Aw_20130426_20130725.csv"
#myFile <- c("QC_test2_Aw_20130426_20130725.csv"
#           , "QC_test2_Aw_20130725_20131015.csv", "QC_test2_Aw_20140901_20140930.csv")
myDir.import <- file.path(getwd(),"Data2_QC")
myDir.export <- file.path(getwd(),"Data4_Stats")
fun.Stats.File(myFile, myDir.import, myDir.export)
}
\keyword{continuous}
\keyword{data,}
\keyword{statistics}
