% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.miniDOT.cat.R
\name{minidot_cat}
\alias{minidot_cat}
\title{Compile DO data}
\usage{
minidot_cat(folderpath, savetofolder, local_tz = Sys.timezone(location = TRUE))
}
\arguments{
\item{folderpath}{The folder where the individual files are located}

\item{savetofolder}{The folder where the output csv is to be saved}

\item{local_tz}{Local time zone used for converting from UTC time zone.
Default is Sys.timezone(location = TRUE).}
}
\value{
a .csv file
}
\description{
This function compiles Dissolved Oxygen data files downloaded
from PME miniDot DO loggers. Use if the data where not compiled using the PME
java app on the logger.
}
\details{
Original function (compile_do_folder) written by Tim Martin.
https://github.com/mnsentinellakes/mnsentinellakes/blob/master/R/compile_do_folder.R

Edits:

* Renamved function

* Added local_tz as a parameter.

* Replaced lubridate::use_tz with as.POSIXct.
}
\examples{
\dontrun{
dn <- tempdir()
compile_do_folder(
  folderpath = file.path(dn, "Greenwood/2019/7450-516388"),
  savetopath = file.path(dn, "Greenwood/2019")
}

}
