% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarify.R
\name{rarify}
\alias{rarify}
\title{Rarify (subsample) biological sample to fixed count}
\usage{
rarify(inbug, sample.ID, abund, subsiz, mySeed = NA)
}
\arguments{
\item{inbug}{Input data frame.  Needs 3 columns (SampleID, taxonomicID, Count).}

\item{sample.ID}{Column name in inbug for sample identifier.}

\item{abund}{Column name in inbug for organism count.}

\item{subsiz}{Target subsample size for each sample.}

\item{mySeed}{Seed for random number generator.  If provided the results with the same inbug file will produce the same results. Defaut = NA (random seed will be used.)}
}
\value{
Returns a data frame with the same three columns but the abund field has been modified so the total count for each sample is no longer above the target (subsiz).
}
\description{
Takes as an input a 3 column data frame (SampleID, TaxonID, Count) and returns a similar dataframe with revised Counts.
}
\details{
The other inputs are subsample size (target number of organisms in each sample) and seed.
The seed is given so the results can be reproduced from the same input file.  If no seed is given a random seed is used.

rarify function:
 R function to rarify (subsample) a macroinvertebrate sample down to a fixed count;
 by John Van Sickle, USEPA. email: VanSickle.John@epa.gov    ;
 Version 1.0, 06/10/05;
}
\examples{
# Subsample to 200 organisms for a single sample from MS.

# load bio data
DF.biodata <- data_bio2rarify
dim(DF.biodata)
View(DF.biodata)

# subsample
mySize <- 200
Seed.MS <- 18171210
bugs.mysize <- rarify(inbug=DF.biodata, sample.ID="SampRep"
                     ,abund="Count",subsiz=mySize, mySeed=Seed.MS)

# view results
dim(bugs.mysize)
View(bugs.mysize)

# Compare pre- and post- subsample counts
df.compare <- merge(DF.biodata, bugs.mysize, by=c("SampRep", "tax"))
View(df.compare)

# save the data
#write.table(bugs.mysize,paste("bugs",mySize,"txt",sep="."),sep="\\t")
}
