% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.Helper.R
\name{rquery.t.test}
\alias{rquery.t.test}
\title{Rquery T-Test}
\usage{
rquery.t.test(x, y = NULL, paired = FALSE, graph = TRUE, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{an optional (non-empty) numeric vector of data values}

\item{paired}{: if TRUE, paired t-test is performed}

\item{graph}{: if TRUE, the distribution of the data is shown for the
inspection of normality}

\item{...}{: further arguments to be passed to the built-in t.test()
R function}
}
\value{

}
\description{
Performs one or two samples t-test
}
\details{
1. shapiro.test is used to check normality
2. F-test is performed to check equality of variances
 If the variances are different, then Welch t-test is used
}
