% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyExample.R
\name{runShinyExample}
\alias{runShinyExample}
\title{run Shiny Example}
\usage{
runShinyExample()
}
\description{
Launches Shiny app.
}
\details{
The Shiny app based on the R package ContDataQC is included in the R package.
This function launches that app.

The Shiny app is online at:
https://davidagibbs.shinyapps.io/rmn_continuous_data_active/
}
\examples{
\dontrun{
# Run Function
runShinyExample()
}
}
