% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.UserDefinedValues.R
\docType{data}
\name{UserDefinedValues}
\alias{UserDefinedValues}
\title{User Defined Values}
\format{An object of class \code{logical} of length 1.}
\usage{
UserDefinedValues
}
\description{
User defined values for variables used across multiple functions in this library.
The user had the ability to modify the values for names, units, QC thresholds, etc.
}
\examples{
#Not intended to be accessed indepedant of function ContDataQC().
#Data values only.  No functions.
}
\keyword{continuous}
\keyword{data}

