% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.CustomConfig.R
\name{config.load}
\alias{config.load}
\title{Custom Configuration}
\usage{
config.load(myFile, myExt = "R")
}
\arguments{
\item{myFile}{Configuration file (including path if not in working directory)}

\item{myExt}{"R" or "RDS"}
}
\value{
NA, values in file added to environment "ContData.env".
}
\description{
Function to load custom configuration of user defined values.
Adds to the existing environment.
User should load a new configuration prior to master script.
}
\details{
Custom config files must **not** include
`ContData.env <- new.env(parent = emptyenv())`.
This will create a 2nd environment that will not be accessed by the functions
in the package.
}
