% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfun.ContDataQC.R
\name{ContDataQC}
\alias{ContDataQC}
\title{Master Script}
\usage{
ContDataQC(fun.myData.Operation, fun.myData.SiteID, fun.myData.Type,
  fun.myData.DateRange.Start, fun.myData.DateRange.End, fun.myDir.BASE,
  fun.myDir.SUB.import, fun.myDir.SUB.export)
}
\description{
Calls other scripts
}

