% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.QC.R
\name{fun.QC}
\alias{fun.QC}
\title{Quality Control}
\usage{
fun.QC(fun.myData.SiteID, fun.myData.Type = "QC", fun.myData.DateRange.Start,
  fun.myData.DateRange.End, fun.myDir.import = "", fun.myDir.export = "",
  fun.myReport.format, fun.myReport.Dir, fun.CreateReport = TRUE)
}
\arguments{
\item{fun.myData.SiteID}{Station/SiteID.}

\item{fun.myData.Type}{data type is "QC".}

\item{fun.myData.DateRange.Start}{Start date for requested data. Format = YYYY-MM-DD.}

\item{fun.myData.DateRange.End}{End date for requested data. Format = YYYY-MM-DD.}

\item{fun.myDir.import}{Directory for import data.  Default is current working directory.}

\item{fun.myDir.export}{Directory for export data.  Default is current working directory.}

\item{fun.myReport.format}{Report format (docx or html).  Default is specified in config.R (docx).  Can be customized in config.R; ContData.env$myReport.Format.}

\item{fun.myReport.Dir}{Report (rmd) template folder.  Default is the package rmd folder.  Can be customized in config.R; ContData.env$myReport.Dir.}

\item{fun.CreateReport}{Boolean parameter to create reports or not.  Default = TRUE.}
}
\value{
Returns a csv file to specified directory with QC flags.
}
\description{
Subfunction for performing QC.  Needs to be called from ContDataQC().
Requires zoo().
}
\examples{
#Not intended to be accessed indepedant of function ContDataQC().
}
