% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.Report.R
\name{fun.Report}
\alias{fun.Report}
\title{Generate Reports}
\usage{
fun.Report(
  fun.myData.SiteID,
  fun.myData.Type,
  fun.myData.DateRange.Start,
  fun.myData.DateRange.End,
  fun.myDir.import = getwd(),
  fun.myDir.export = getwd(),
  fun.myFile.Prefix,
  fun.myReport.format,
  fun.myReport.Dir
)
}
\arguments{
\item{fun.myData.SiteID}{Station/SiteID.}

\item{fun.myData.Type}{data type;
c("Air","Water","AW","Gage","AWG","AG","WG")}

\item{fun.myData.DateRange.Start}{Start date for requested data.
Format = YYYY-MM-DD.}

\item{fun.myData.DateRange.End}{End date for requested data.
Format = YYYY-MM-DD.}

\item{fun.myDir.import}{Directory for import data.
Default is current working directory.}

\item{fun.myDir.export}{Directory for export data.
Default is current working directory.}

\item{fun.myFile.Prefix}{Valid prefixes are "QC", "DATA", or "STATS".
This determines the RMD to use for the output.}

\item{fun.myReport.format}{Report format (docx or html).
Default is specified in config.R (docx).  Can be customized in config.R;
ContData.env$myReport.Format.}

\item{fun.myReport.Dir}{Report (rmd) template folder.
Default is the package rmd folder.  Can be customized in config.R;
ContData.env$myReport.Dir.}
}
\value{
Creates a Word file in the specified directory.
}
\description{
Subfunction for generating reports.  Needs to be called from ContDataQC().
Requires knitr() and pandoc.
Different reports (in Word) are created using the RMD files in this package
depending on the input fun.myFile.Prefix.
Values are "QC", "DATA", or "STATS".
}
\examples{
#Not intended to be accessed indepedant of function ContDataQC().
}
