% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.miniDOT.cat.R
\name{minidot_cat}
\alias{minidot_cat}
\title{Compile DO data}
\usage{
minidot_cat(
  folderpath,
  savetofolder,
  filename = NULL,
  local_tz = Sys.timezone(location = TRUE)
)
}
\arguments{
\item{folderpath}{The folder where the individual files are located}

\item{savetofolder}{The folder where the output csv is to be saved}

\item{filename}{Output filename.  Default = SerialNumber.csv}

\item{local_tz}{Local time zone used for converting from UTC time zone.
Default is Sys.timezone(location = TRUE).}
}
\value{
In ouput folder a CSV file
}
\description{
This function compiles Dissolved Oxygen data files downloaded
from PME miniDot DO loggers. Use if the data where not compiled using the PME
java app on the logger.
}
\details{
Original function (compile_do_folder) written by Tim Martin.

https://github.com/mnsentinellakes/mnsentinellakes/blob/master/R/compile_do_folder.R

Edits:

* Renamed function

* Added local_tz as a parameter.

* Replaced lubridate::use_tz with as.POSIXct.

* Added filename parameter.

* Swaped order of operations so fail condition is triggered it happens before
any work is done.

* Added working example with data.
}
\examples{
# Data
dn_input <- file.path(system.file("extdata", package = "ContDataQC")
                      , "minidot")
dn_export <- tempdir()

# Function
minidot_cat(folderpath = dn_input, savetofolder = dn_export)

}
